Create ATP Assays Market.md
ATP is a universal mediator of metabolism and signaling across unicellular and multicellular species. There is a basic interdependence between the dynamics of ATP and the physiology that arises at the intracellular and extracellular levels. Thus, characterizing and recognizing ATP dynamics offers valuable mechanistic insights into processes that varies from neurotransmission to the chemotaxis of immune cells. Therefore, there is a requirement for a methodology to interrogate both the temporal and spatial components of ATP dynamics from the subcellular to organismal levels in live specimens