Create Blood Screening Market.md
With the increasing number of blood donations, awareness of blood verification & safety, and spending on healthcare around the world, more advanced technologies are widely used in high income countries and will be widely adopted in middle and low-income countries within the next 10 years. However, most developing countries, including India and China currently use ELISA. The growth of the market for blood screening for NAT is severely inhibited in the present timeframe.